Create License
License for all content of this repository. Intention is for anyone to be able to take this content, adapt and modify as they see fit provided they attribute the origination to the First Church of the Manda. This includes the sale of adapted and derivative physical works, provided that digital version are freely available.